consumer/requirements.txt
psycopg2-binary~=2.8.6
paho-mqtt==1.6.1
ipdb